菜单展示，添加菜品 over

+++ controllers/DishController.php
@@ -7,20 +7,24 @@
  */
 namespace app\controllers;
 
+use app\models\LoginForm;
 use Yii;
 use yii\base\Controller;
 use yii\data\Pagination;
 use yii\db\Query;
 use app\models\Dish;
+use yii\web\NotFoundHttpException;
 
-class DishController extends Controller{
+class DishController extends Controller
+{
 
 
     /**
      * @return string
      * 添加菜单
      */
-    public function actionAdd(){
+    public function actionAdd()
+    {
         $dish = new Dish();
         $res = Yii::$app->request;
         $dish->s_id = Yii::$app->user->id;
@@ -31,26 +35,32 @@ class DishController extends Controller{
         $dish->desc = $res->post('desc');
         $dish->p_id = $res->post('p_id');
         $dish->sale_num = 0;
-        if($dish->validate()&&$dish->save()){
-            return 'success';
-        }else{
-            return 'fales';
+        if ($dish->validate() && $dish->save()) {
+            return $this->render('add',['model'=>$dish]);
+        } else {
+            return $this->render('add',['model'=>$dish]);
         }
     }
 
     /**
      *删除 传入菜品id
      */
-    public function actionDel(){
-        $id = Yii::$app->request->post('id');
-        $s_id = Yii::$app->user->id;
-        Dish::find()->where(['s_id'=>$s_id,'id'=>$id])->one()->delete();
+    public function actionDel()
+    {
+        if (Yii::$app->request->isAjax) {
+            $id = Yii::$app->request->post('id');
+            $s_id = Yii::$app->user->id;
+            Dish::find()->where(['s_id' => $s_id, 'id' => $id])->one()->delete();
+            return Yii::$app->response->redirect(['dish/dishplay']);
+        }
     }
+
 
     /**
      *修改 传入菜品id还有需要修改的信息
      */
-    public function actionUpdate(){
+    public function actionUpdate()
+    {
 
         $id = Yii::$app->request->post('dish_id');
         $res = Yii::$app->request;
@@ -71,15 +81,16 @@ class DishController extends Controller{
         }
     }
 
-    public function actionDisplay(){
+    public function actionDisplay()
+    {
 
         $store_id = Yii::$app->user->id;
 
         $query = (new Query())->from('dish')
-            ->where(['s_id' =>$store_id]);
+            ->where(['s_id' => $store_id]);
 
         $pagination = new Pagination([
-            'defaultPageSize' => 2,
+            'defaultPageSize' => 8,
             'totalCount' => $query->count(),
         ]);
 
@@ -88,6 +99,6 @@ class DishController extends Controller{
             ->limit($pagination->limit)
             ->all();
 
-        return $this->render('display',['list'=>$res,'page' => $pagination]);
+        return $this->render('display', ['list' => $res, 'page' => $pagination]);
     }
 }
